Started shifting alert functions away from DangerAlertCore
Added AlertBase.cs, CollisionAlert.cs, and ResourceAlert.cs, and added a
foreach loop, and Alerts list to DangerAlertCore, instead of calling
individual functions. This will make it easier to implement low resource
alerts and the like. TODO: Make the new GUI, and restore old
functionality of the GUI, including persistence.